Ignoring invalid IPv4s in IP filter
To reach a bit more consistency with the Tshock IP filter, I implemented here as well that if one of the blocks of the detected IPv4-structured substring has a value of over 255 then that substring would be allowed by the filter as that is not a valid IPv4 (93.420.69.666 is allowed as that is not a valid IPv4, 127.0.0.1 is not). 
Also divided the checks against IPv4s and known server link ips because the two checks now require different approaches between eachother.